+++ thesis-gui/src/components/category/HomePage.tsx
@@ -1,36 +1,20 @@
-import React, {useEffect, useState} from "react";
-import {useCategoryContext} from "../../context/CategoryContext";
-import {useQuizzContext} from "../../context/QuizzContext";
-
-import {useHistory} from "react-router";
+import React, { useEffect } from "react";
 import {
-    IonButton,
+    IonCard, IonCardContent, IonCardSubtitle, IonCardTitle,
     IonContent,
-    IonInput,
-    IonItem,
-    IonList,
     IonLoading,
     IonPage,
-    IonRadioGroup,
-    IonText
+    IonRouterLink
 } from "@ionic/react";
-import {CategoryCheckbox} from "./CategoryCheckbox";
+import {useHistory} from "react-router";
+import {useCategoryContext} from "../../context/CategoryContext";
 import {PageHeader} from "../common/PageHeader";
 
+const PAGE_NAME = 'Play';
+
 export const HomePage: React.FC = () => {
-    const { loadingState, categories } = useCategoryContext();
-    const { chooseKey } = useQuizzContext();
+    const { loadingState } = useCategoryContext();
     const history = useHistory();
-    const [input, setInput] = useState<string>('');
-    const [disableQuizz, setDisableQuizz] = useState<boolean>(true);
-    const handleCodeLength = (code: string) => {
-        if(code.length != 8)
-            setDisableQuizz(true);
-        else
-            setDisableQuizz(false);
-
-    };
-
 
     useEffect(() => {
         if (loadingState === 'FAILURE') {
@@ -38,56 +22,39 @@ export const HomePage: React.FC = () => {
         }
     }, [loadingState, history])
 
-    const categoryItems = categories.map(category =>
-        <CategoryCheckbox category={category} key={category.name}/>
-    );
-
     return (
         <IonPage>
-            <PageHeader name="Home Page" condense={ false } />
+            <PageHeader name={ PAGE_NAME } />
             <IonContent class="ion-padding">
-                <PageHeader name="Home Page" condense={ true } />
-                <IonLoading isOpen={loadingState === 'LOADING'}/>
-                <IonText color="primary">
-                    <h1>Private Quiz</h1>
-                </IonText>
-                <IonItem>
-                    <IonInput 
-                    value={input}
-                    onIonChange={(e:any) => {
-                        setInput(e.target.value);
-                        chooseKey(e.target.value);
-                        handleCodeLength(e.target.value);
-                    }}
-                    placeholder="Enter code of the Quiz"
-                    ></IonInput>
-                </IonItem>
-                <IonButton
-                    routerLink={"/quizz"}
-                    routerDirection="back"
-                    expand="block"
-                    disabled={disableQuizz}
-                    style={{ marginTop: 20, marginBottom: 30 }}
-                >
-                    Start
-                </IonButton>
-                
-                <IonText color="primary">
-                    <h1>Categories</h1>
-                </IonText>
-                <IonList>
-                    <IonRadioGroup>
-                        { categoryItems }
-                    </IonRadioGroup>
-                </IonList>
-                <IonButton
-                    routerLink="/questions"
-                    routerDirection="back"
-                    expand="block"
-                    style={{ marginTop: 20, marginBottom: 30 }}
-                >
-                    Choose
-                </IonButton>
+                <IonLoading isOpen={loadingState === 'LOADING'} />
+
+                <IonCard>
+                    <IonCardContent>
+                        <IonCardTitle style={{ textAlign: "center" }}>
+                            Private Quiz
+                        </IonCardTitle>
+                        <IonCardSubtitle style={{ textAlign: "center" }}>
+                            Play to earn points and climb the leaderboard!
+                        </IonCardSubtitle>
+                        <IonCardSubtitle style={{ textAlign: "center" }}>
+                            Click <IonRouterLink routerLink="/private-quiz">here</IonRouterLink> to start the game
+                        </IonCardSubtitle>
+                    </IonCardContent>
+                </IonCard>
+
+                <IonCard>
+                    <IonCardContent>
+                        <IonCardTitle style={{ textAlign: "center" }}>
+                            Categories
+                        </IonCardTitle>
+                        <IonCardSubtitle style={{ textAlign: "center" }}>
+                            Play to earn points and climb the leaderboard!
+                        </IonCardSubtitle>
+                        <IonCardSubtitle style={{ textAlign: "center" }}>
+                            Click <IonRouterLink routerLink="/categories">here</IonRouterLink> to start the game
+                        </IonCardSubtitle>
+                    </IonCardContent>
+                </IonCard>
             </IonContent>
         </IonPage>
     );
