--- thesis-gui/src/types/teacher-question-set.tsx
@@ -1,6 +1,0 @@
-import { TeacherQuestion } from "./teacher-question"
-
-export type TeacherQuestionSet = {
-    questionsSetName: string,
-    teacherQuestionsResponse: TeacherQuestion[],
-}
