--- thesis-gui/src/types/teacher-question.tsx
@@ -1,5 +1,0 @@
-export type TeacherQuestion = {
-    question: string,
-    correct: string,
-    answers: string[]
-}
